requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.4.0
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2==2.8.5
whitenoise==5.2.0
DateTime==4.2.0
ics==0.7
requests==2.24.0
django-oauth-toolkit==1.3.3
google-api-python-client==1.12.5
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2